move Next.js types to interfaces.ts
